pyproject.toml
[project]
name = "nexus-pos"
version = "1.0.0"
description = "Sistema POS Multi-Tenant con FastAPI, SQLModel y PostgreSQL"
authors = [
    {name = "Nexus POS Team"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # FastAPI Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    
    # Database
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    
    # Security & Authentication
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.2",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0",
    "requests>=2.31.0",
    
    # Payment Integration
    "mercadopago>=2.2.1",
    
    # Resilience & Monitoring
    "tenacity>=8.2.3",
    "sentry-sdk[fastapi]>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.9",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]

afip = [
    # Descomentar cuando se implemente AFIP
    # "pyafipws>=2.9.0",
]

celery = [
    # Para tareas de fondo con Celery
    "celery[redis]>=5.3.4",
    "redis>=5.0.1",
    "flower>=2.0.1",
]

apscheduler = [
    # Alternativa a Celery para tareas programadas
    "apscheduler>=3.10.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=app --cov-report=term-missing --cov-report=html"

[dependency-groups]
dev = [
    "ruff>=0.1.9",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]
